moved DiffElement to UIElements.js file
